% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FarmerB_FarmName_bbox.R
\docType{data}
\name{FarmerB_FarmName_bbox}
\alias{FarmerB_FarmName_bbox}
\title{FarmerB_FarmName_bbox}
\format{
A data frame with 1 rows and 2 variables:
\describe{
\item{\code{id}}{double ID of boundary.}
\item{\code{geometry}}{list Coordinates of bounding box polygon.}
}
}
\usage{
FarmerB_FarmName_bbox
}
\description{
Bounding box of the farm that encompasses Farmer B's
experimental OFPE fields.
}
\keyword{datasets}
