% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FarmerB_pro_2016.R
\docType{data}
\name{FarmerB_pro_2016}
\alias{FarmerB_pro_2016}
\title{FarmerB_pro_2016}
\format{
A data frame with 1183 rows and 58 variables:
\describe{
\item{\code{sample_id}}{character Field name.}
\item{\code{sp1}}{double Infrared spectrometer measurement.}
\item{\code{sp2}}{double Infrared spectrometer measurement.}
\item{\code{sp3}}{double Infrared spectrometer measurement.}
\item{\code{sp4}}{double Infrared spectrometer measurement.}
\item{\code{sp5}}{double Infrared spectrometer measurement.}
\item{\code{sp6}}{double Infrared spectrometer measurement.}
\item{\code{sp7}}{double Infrared spectrometer measurement.}
\item{\code{sp8}}{double Infrared spectrometer measurement.}
\item{\code{sp9}}{double Infrared spectrometer measurement.}
\item{\code{sp10}}{double Infrared spectrometer measurement.}
\item{\code{sp11}}{double Infrared spectrometer measurement.}
\item{\code{sp12}}{double Infrared spectrometer measurement.}
\item{\code{sp13}}{double Infrared spectrometer measurement.}
\item{\code{sp14}}{double Infrared spectrometer measurement.}
\item{\code{sp15}}{double Infrared spectrometer measurement.}
\item{\code{sp16}}{double Infrared spectrometer measurement.}
\item{\code{sp17}}{double Infrared spectrometer measurement.}
\item{\code{sp18}}{double Infrared spectrometer measurement.}
\item{\code{sp19}}{double Infrared spectrometer measurement.}
\item{\code{sp20}}{double Infrared spectrometer measurement.}
\item{\code{sp21}}{double Infrared spectrometer measurement.}
\item{\code{sp22}}{double Infrared spectrometer measurement.}
\item{\code{sp23}}{double Infrared spectrometer measurement.}
\item{\code{sp24}}{double Infrared spectrometer measurement.}
\item{\code{sp25}}{double Infrared spectrometer measurement.}
\item{\code{sp26}}{double Infrared spectrometer measurement.}
\item{\code{sp27}}{double Infrared spectrometer measurement.}
\item{\code{sp28}}{double Infrared spectrometer measurement.}
\item{\code{sp29}}{double Infrared spectrometer measurement.}
\item{\code{sp30}}{double Infrared spectrometer measurement.}
\item{\code{sp31}}{double Infrared spectrometer measurement.}
\item{\code{sp32}}{double Infrared spectrometer measurement.}
\item{\code{sp33}}{double Infrared spectrometer measurement.}
\item{\code{sp34}}{double Infrared spectrometer measurement.}
\item{\code{sp35}}{double Infrared spectrometer measurement.}
\item{\code{sp36}}{double Infrared spectrometer measurement.}
\item{\code{sp37}}{double Infrared spectrometer measurement.}
\item{\code{sp38}}{double Infrared spectrometer measurement.}
\item{\code{protein}}{double Grain protein content (\%).}
\item{\code{moisture}}{double Moisture content(\%).}
\item{\code{slope1}}{logical Protein calibration/calculation slope.}
\item{\code{bias1}}{double Protein calibration/calculation bias.}
\item{\code{slope2}}{double Moisture calibration/calculation slope.}
\item{\code{bias2}}{integer Moisture calibration/calculation bias.}
\item{\code{product}}{character Crop.}
\item{\code{date}}{character Date of harvest.}
\item{\code{time}}{character Time observation was collected.}
\item{\code{int_time}}{integer Time variable in integer form.}
\item{\code{temp}}{logical Ambient temperature.}
\item{\code{gps}}{character GPS code.}
\item{\code{utc_time}}{integer Coordinated Universal Time (UTC).}
\item{\code{y}}{double Latitude in decimal degrees.}
\item{\code{n}}{character Northern hemisphere indicator.}
\item{\code{x}}{double Longitude in decimal degrees.}
\item{\code{w}}{character Western hemisphere indicator.}
\item{\code{id}}{integer ID column for each observation.}
\item{\code{orig_file}}{character Original file name.}
}
}
\usage{
FarmerB_pro_2016
}
\description{
Grain protein content (\%) data collected for winter wheat from
Farmer B's CropScan 3300H infrared spectrometer in 2016. This product is
built by NextInstruments. Infrared scanner is mounted in the cab of the
combine to maximize data quality and is automatically collected ~10 seconds
during harvest.
}
\details{
DETAILS
}
\keyword{datasets}
