% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FarmerC_SR_2019.R
\docType{data}
\name{FarmerC_SR_2019}
\alias{FarmerC_SR_2019}
\title{FarmerC_SR_2019}
\format{
A data frame with 23047 rows and 15 variables:
\describe{
\item{\code{Time}}{character Date and time of observation.}
\item{\code{Crop}}{integer Crop code.}
\item{\code{AppldRt}}{double Variable rate of seed applied (lbs/ac).}
\item{\code{SWATHWI}}{double Width of swath.}
\item{\code{DISTANC}}{double Distance between observation points.}
\item{\code{Heading}}{double Direction or bearing.}
\item{\code{SECTION}}{integer Section ID of seeder.}
\item{\code{CntrlRt}}{double Control rate setting.}
\item{\code{TargtRt}}{double Target rate setting.}
\item{\code{Variety}}{character Crop variety.}
\item{\code{Elevatn}}{double Elevation of observation.}
\item{\code{IsoTime}}{character International Organization for Standardization (ISO) date and time format.}
\item{\code{Machine}}{integer Machine identifier code.}
\item{\code{orig_fl}}{character Original shapefile name.}
\item{\code{geometry}}{list Coordinates of seeding points.}
}
}
\usage{
FarmerC_SR_2019
}
\description{
Variable rate seeding rate point data from Farmer C's
John Deere seeder in 2019. Product is spring wheat seeding rate in
pounds of seed per acre. Data is raw experimental data with attributes
recorded by Farmer C.
}
\details{
DETAILS
}
\keyword{datasets}
