% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FarmerC_yld_2019.R
\docType{data}
\name{FarmerC_yld_2019}
\alias{FarmerC_yld_2019}
\title{FarmerC_yld_2019}
\format{
A data frame with 258411 rows and 15 variables:
\describe{
\item{\code{DISTANCE}}{double Distance between observation points}
\item{\code{SWATHWIDTH}}{double Width of swath.}
\item{\code{VRYIELDVOL}}{double Crop yield in bushels per acre.}
\item{\code{SECTIONID}}{integer Section identifier.}
\item{\code{Crop}}{integer Crop code.}
\item{\code{WetMass}}{double Mass of wet product.}
\item{\code{Moisture}}{double Moisture content (\%).}
\item{\code{Time}}{character Date and time of observation.}
\item{\code{Heading}}{double Direction or bearing.}
\item{\code{VARIETY}}{character Crop variety.}
\item{\code{Elevation}}{double Elevation of observation point.}
\item{\code{IsoTime}}{character International Organization for Standardization (ISO) date and time format.}
\item{\code{Machine}}{integer Machine identifier code.}
\item{\code{orig_file}}{character Original file name of shapefile.}
\item{\code{geometry}}{list Coordinates of harvest points.}
}
}
\usage{
FarmerC_yld_2019
}
\description{
Spring wheat crop yield (bushels per acre) data
collected from Farmer C's John Deere Combine in 2019. Yield
data is automatically collected ~3 seconds during harvest.
}
\details{
DETAILS
}
\keyword{datasets}
