% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/henrys_bbox.R
\docType{data}
\name{henrys_bbox}
\alias{henrys_bbox}
\title{henrys_bbox}
\format{
A data frame with 1 rows and 2 variables:
\describe{
\item{\code{id}}{double ID of boundary.}
\item{\code{geometry}}{list Coordinates of bounding box polygon.}
}
}
\usage{
henrys_bbox
}
\description{
Field boundary for Farmer C's OFPE experimental field
named 'henrys'.
}
\details{
DETAILS
}
\keyword{datasets}
