% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/millview_bbox.R
\docType{data}
\name{millview_bbox}
\alias{millview_bbox}
\title{millview_bbox}
\format{
A data frame with 1 rows and 2 variables:
\describe{
\item{\code{id}}{double ID of boundary.}
\item{\code{geometry}}{list Coordinates of bounding box polygon.}
}
}
\usage{
millview_bbox
}
\description{
Field boundary for Farmer C's OFPE experimental field
named 'millview'.
}
\details{
DETAILS
}
\keyword{datasets}
