% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sec35middle_bbox.R
\docType{data}
\name{sec35middle_bbox}
\alias{sec35middle_bbox}
\title{sec35middle_bbox}
\format{
A data frame with 2 rows and 2 variables:
\describe{
\item{\code{id}}{double ID of boundary.}
\item{\code{geometry}}{list Coordinates of bounding box polygon.}
}
}
\usage{
sec35middle_bbox
}
\description{
Field boundary for Farmer B's OFPE experimental field
named 'sec35mid'.
}
\details{
DETAILS
}
\keyword{datasets}
