% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unzipShapefiles.R
\name{unzipShapefiles}
\alias{unzipShapefiles}
\title{unzipShapefiles}
\usage{
unzipShapefiles(zip_file, path)
}
\arguments{
\item{zip_file}{character}

\item{path}{character}
}
\value{
sf
}
\description{
Unzips compressed files containing ESRI shapefiles from the
'data-raw' folder of the OFPEDATA package. Takes a zipped file
name and returns a 'sf' object.
}
\examples{
\dontrun{
unzipShapefiles("zipped_data_filename", "path_to_data")
}

}
